Clear out linked list and add append and toString
The original idea was to go in to the project with an array, which
defeats the purpose of creating a linked list. After realizing this
error, I cleared out the linked list class.

The append method was added to create new nodes and add them to the end
of the linked list with pointers to each object.

The toString method was created at the same time to give some output
to ensure the append method as well as future methods are created
correctly.